fix(back)migration

+++ back/migrations/revert/remove_value_func.sql
@@ -2,7 +2,7 @@
 
 BEGIN;
 	
-CREATE FUNCTION wallet_value(IN x real, IN y int, OUT product int)
+CREATE FUNCTION wallet_value(IN x real, IN y int, OUT product DOUBLE PRECISION)
 	AS 'SELECT x * y'
 	LANGUAGE SQL;
